feat: Make emily run as local warp service instead of apigateway.
