+++ emily/handler/Cargo.toml
@@ -4,7 +4,11 @@ version = "0.1.0"
 edition = "2021"
 
 [[bin]]
-name = "emily-handler"
+name = "emily-lambda"
+
+[[bin]]
+name = "emily-server"
+features = ["testing"]
 
 [features]
 default = ["testing"]
