+++ emily/handler/src/lib.rs
@@ -5,3 +5,4 @@ pub mod api;
 pub mod common;
 pub mod context;
 pub mod database;
+pub mod logging;
